+++ semi/src/semi/customer/CustomerFindServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package semi.customer;
 
 import java.io.IOException;
 
@@ -10,8 +10,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import beans.CustomerDao;
 
-@WebServlet(urlPatterns = "/customer/change_pw.do")
-public class CustomerChangePwServlet extends HttpServlet{
+@WebServlet(urlPatterns = "/customer/find.do")
+public class CustomerFindServlet extends HttpServlet{
 
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -20,20 +20,25 @@ public class CustomerChangePwServlet extends HttpServlet{
 			
 			req.setCharacterEncoding("UTF-8");
 			
-			String customer_id = (String)req.getSession().getAttribute("customer_id");
-			String customer_pw = req.getParameter("customer_pw");
+			String customer_name= req.getParameter("customer_name");
+			String customer_phone = req.getParameter("customer_phone");
 			
 			CustomerDao dao = new CustomerDao();
-//			dao.changePw(customer_pw);
+			String customer_id = dao.find(customer_name, customer_phone);
 			
+			if(customer_id == null) {
+				resp.sendRedirect("find_result.jsp");
+			}
+			else {
+				resp.sendRedirect("find_result.jsp?customer_id=" + customer_id);
+			}
 			
-			
+					
 		} catch (Exception e) {
 			e.printStackTrace();
 			resp.sendError(500);
 		}
-		
-		
+	
 	}
 	
 }
