+++ semi/src/semi/customer/CustomerLoginServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package semi.customer;
 
 import java.io.IOException;
